Refactor: Extract html module and remove duplicated code
- Create html.rs with shared render_template_with_components function
- Remove duplicate rendering logic from lib.rs and server.rs
- Add helper function to reduce code duplication in server.rs
- Add unit tests for html module
